Added UART communication and NMEA protocol
Added UART.c and UART.h which communicate to a terminal with UART. message_temperature is temporarily using normal UART while message_clock uses NMEA protocol adapted UART.

Added NMEA.c and NMEA.h. They allow for UART to adapt an NMEA like protocol. I could not adapt the full protocol since its written for CAN but this is a version of it which I'll update to perform better in the future.